Updated ode45 movie making and fixed eigs movie making so that it works.
ppOde45.makeMovie() now increments the colorbar scale only when the
scale range is either halved or doubled.  The maxfactor crop is still
used.

ppEigs.makeMovie() requires that you enter the name of the 'evals'
file that you want to process.  Number of periods and frame per period
over which to plot are entered in the POSTPPARS.dat file.